skyobejct/ move the Incs back from the registry. Move Split core to regsitry/: still draft
